+++ .gitignore
@@ -1,0 +1,2 @@
+go.fitbit
+example/token
